% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Get.R
\name{GetNumericVector}
\alias{GetCharacterVector}
\alias{GetLogicalVector}
\alias{GetNumericVector}
\title{Generate random vectors}
\usage{
GetNumericVector(seed = 1)

GetLogicalVector(seed = 1)

GetCharacterVector(seed = 1)
}
\arguments{
\item{seed}{A positive interger for the random number generator. Default 1.}
}
\description{
These functions create vectors of numerics, logicals or characters.
Given a seed, the functions will always return the
same subset.
}

