% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nclRintroduction-package.R
\docType{package}
\name{nclRintroduction-package}
\alias{nclRintroduction}
\alias{nclRintroduction-package}
\alias{nclrintroduction}
\title{R courses at Newcastle University}
\description{
Functions and datasets used in for R
courses at Newcastle University. This package is used in
the "Introduction to R" course".
}
\keyword{package}

